Refactor
The "genericserver" has been split into the following modules:
- cmd (contains functions to be easily called from cli)
- serve (contains functions related to serving endpoints via http wiht gin-going)
- handlers (contains generic http handlers)
- config (contains a modular configuration)

About config

Viper dependency has been removed in order to allow using custom types with
ease.  For example, a configuration now can set a common.Address type, and the
hex decoding is done automatically.  Instead of yaml, the config now uses TOML.

Instead of storing the config in a global variable called C, the config is just
unmarshalled in a local variable.  Using the `validate:"required"` tags allows
forcing a configuration option be exist in the config file.

New configuration examples need to be generated for the different servers.
Currently, the test vectors in `config/config_test.go` can be used as a guide.

About servers

The servers are now in the folder "servers" insetad of "cmd", with some generic
functions moved outside.  Also, cli tools that were not related to servers have
been moved to "bintools".

About Identity

Since most of the servers have an identity, a new struct has been defined that
collects many services, components and objects that usually go together with an
identity in a server.  This struct is `loaders.Identity`, and is commonly
passed to the handlers, removing the usage of global variables that hold
services.  In order to have gingonic handlers that take more arguments than
just the gingonic Context, an addaptor function has been added: `WithIden`.
This pattern allows removing the usage of global variables and removes the need
for repeated code.

A generalization of the Identity object may be needed in the future, so it is
open to improvement.

NOTE: The tests pass, but I haven't tried running the servers, this will be
required next step.

NOTE: There are no global variables now.  Please, keep it this way!
