+++ cmd/post.go
@@ -1,0 +1,30 @@
+package cmd
+
+import (
+	"bytes"
+	"io/ioutil"
+	"net/http"
+	"strings"
+)
+
+func PostAdminApi(addr string, command string) (string, error) {
+	hostport := strings.Split(addr, ":")
+	if hostport[0] == "0.0.0.0" {
+		hostport[0] = "127.0.0.1"
+	}
+	url := "http://" + hostport[0] + ":" + hostport[1] + "/" + command
+
+	var body bytes.Buffer
+	resp, err := http.Post(url, "text/plain", &body)
+	if err != nil {
+		return "", err
+	}
+
+	defer resp.Body.Close()
+	output, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return "", err
+	}
+
+	return string(output), nil
+}
