--- cmd/relay/.gitignore
@@ -1,2 +1,0 @@
-private-configs
-r.sh
